Move sql queries off main thread
